fix(customer): rename file dto to create customer
